+++ xous-rs/src/definitions/messages/id.rs
@@ -1,0 +1,1 @@
+pub type Id = usize;
