--- xous-rs/src/messages/mod.rs
@@ -1,1 +1,0 @@
-
